28-CHNAGE: element file restructure
