+++ apps/meally.com.au/tests/components/Addbatch.test.tsx
@@ -1,5 +1,5 @@
 import { render, fireEvent, screen } from '@testing-library/react';
-import AddBatch from '../../src/common/components/elements/recipe_elemnts/Addbatch';
+import AddBatch from '../../src/common/components/elements/recipe_page/Addbatch';
 
 describe('AddBatch', () => {
   it('Plus in document', async () => {
